Move modules to new package geoanalytics: * evaluation.clustering_ground_truth -> geoanalytics.clustering_ground_truth * util.coordinates -> geoanalytics.coordinates * util.gometry -> geoanalytics.geometry
