+++ src/sensai/geoanalytics/clustering_ground_truth.py
@@ -9,7 +9,7 @@ from geopandas import GeoDataFrame
 from shapely.geometry import Polygon, MultiPoint, MultiPolygon
 from typing import Sequence, Union, Optional
 
-from ..util.coordinates import extractCoordinatesArray, TCoordinates, GeoDataFrameWrapper
+from sensai.geoanalytics.coordinates import extractCoordinatesArray, TCoordinates, GeoDataFrameWrapper
 
 log = logging.getLogger(__name__)
